+++ src/Ui/Bundle/AdminBundle/Form/Type/Event/ExtraParameter/CreateType.php
@@ -13,6 +13,7 @@ namespace Proximum\Vimeet\Ui\Bundle\AdminBundle\Form\Type\Event\ExtraParameter;
 use Proximum\Vimeet\Domain\Event\ExtraParameter\Type;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
+use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 
@@ -36,7 +37,7 @@ class CreateType extends AbstractType
             ->add('name', TextType::class, [
                 'required' => true,
             ])
-            ->add('value', TextType::class, [
+            ->add('value', TextareaType::class, [
                 'required' => true,
             ])
         ;